requirements.txt
pytest==8.3.2
openai==1.0.0
python-dotenv==1.0.0